requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.7
idna==3.11
psycopg==3.2.2
python-dotenv==1.1.1
requests==2.32.5
typing_extensions==4.12.2
urllib3==2.5.0
